refactor: separate Deck DTO mappings
