+++ src/entities/deck/model/rules.ts
@@ -1,7 +1,7 @@
 import type { Card } from "@/entities/card/@x/deck";
 import { createStudyProgressFromCard, isStudyProgressEligible } from "@/entities/study-progress/@x/deck";
 
-import type { Category, Deck, DeckId, DeckStore, DeckView } from "./types";
+import type { Category, DeckDomain, DeckId } from "./types";
 
 const APPLICATION_CATEGORIES: Category[] = ["raw", "math"];
 
@@ -52,24 +52,7 @@ export const getCategory = (category: Category, tags: string[]): Category => {
   return tagCategory ?? category;
 };
 
-/** Keep persistence metadata and mutable store collections behind the Entity boundary. */
-export const toDeckView = (deck: DeckStore): DeckView => ({
-  id: deck.id,
-  name: deck.name,
-  ...(deck.url === undefined ? {} : { url: deck.url }),
-  isPublic: deck.isPublic,
-  scoreMax: deck.scoreMax,
-  scoreMin: deck.scoreMin,
-  selectedTags: [...deck.selectedTags],
-  tagAndFilter: deck.tagAndFilter,
-  category: deck.category,
-  convertToBr: deck.convertToBr,
-  createdAt: deck.createdAt,
-  updatedAt: deck.updatedAt,
-  localMode: deck.localMode,
-});
-
-const isCardMatchingTags = (card: Card, deck: Pick<Deck, "selectedTags" | "tagAndFilter">) => {
+const isCardMatchingTags = (card: Card, deck: Pick<DeckDomain, "selectedTags" | "tagAndFilter">) => {
   const tags = deck.selectedTags;
   if (tags.length === 0) return true;
   if (deck.tagAndFilter) return tags.every((tag) => card.tags.includes(tag));
@@ -78,7 +61,7 @@ const isCardMatchingTags = (card: Card, deck: Pick<Deck, "selectedTags" | "tagAn
 
 export const filterCardsForDeck = <TCard extends Card>(
   cards: TCard[],
-  deck: Pick<Deck, "selectedTags" | "tagAndFilter" | "scoreMax" | "scoreMin">,
+  deck: Pick<DeckDomain, "selectedTags" | "tagAndFilter" | "scoreMax" | "scoreMin">,
   study: { useCardInterval: boolean },
   now: number
 ): TCard[] =>
@@ -95,7 +78,7 @@ export const filterCardsForDeck = <TCard extends Card>(
     );
   });
 
-export const mustFindDeckById = (decks: readonly Deck[], id: DeckId): Deck => {
+export const mustFindDeckById = <TDeck extends DeckDomain>(decks: readonly TDeck[], id: DeckId): TDeck => {
   const deck = decks.find((candidate) => candidate.id === id);
 
   if (deck == null) throw new Error(`Deck not found: ${id}`);
